requirements.txt
Flask
requests
cryptography
python-dotenv